Abstracted Item Equipment - Items can be created via adding in itemcode.txt and corresponding data files. Generated item data files for existing items. Minor UI improvements in Upgrade screen. Corrected Inventory Critical Hit Damage number. Fixed blocking skill queue

+++ TowerConquestServer/src/blockfighter/server/LogicModule.java
@@ -44,7 +44,7 @@ public class LogicModule extends Thread {
     private final ConcurrentLinkedQueue<Projectile> projAddQueue = new ConcurrentLinkedQueue<>();
 
     private long lastRefreshAll = 0;
-    private double lastUpdateTime = 0, lastProcessQueue = 0;
+    private long lastUpdateTime = 0, lastProcessQueue = 0;
 
     private static final ExecutorService LOGIC_THREAD_POOL = Executors.newFixedThreadPool(Globals.SERVER_LOGIC_THREADS,
             new BasicThreadFactory.Builder()
